requirements.txt
# =============================================================
# Dreamscape Mapper — Requirements
# Python 3.8.18 | CUDA 12.x | RTX 4060 Ti
# =============================================================

# ── PyTorch (CUDA 12.1 wheel — last version supporting Py3.8)
# Installed separately via pip index (see setup commands below)
# torch==2.1.2
# torchvision==0.16.2

# ── Core NLP ─────────────────────────────────────────────────
spacy==3.7.4
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl

# ── CRF layer for NER head ────────────────────────────────────
torchcrf==1.1.0

# ── NRC Emotion Lexicon (replaces GoEmotions) ─────────────────
nrclex==3.0.0
# nrclex depends on these — pinned for Py3.8 stability
nltk==3.8.1
textblob==0.17.1

# ── Data loading ─────────────────────────────────────────────
pandas==2.0.3
pyarrow==14.0.2
datasets==2.16.1          # HuggingFace datasets for DreamBank parquet

# ── Scientific stack ──────────────────────────────────────────
numpy==1.24.4             # last version with full Py3.8 support
scikit-learn==1.3.2

# ── Utilities ────────────────────────────────────────────────
tqdm==4.66.2
requests==2.31.0

# ── API Server ───────────────────────────────────────────────
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9